@@ -16,7 +16,7 @@
     "url": "git+https://github.com/Lucifier129/create-app.git"
   },
   "keywords": [
-    "Isomorphism",
+    "isomorphic",
     "router",
     "controller",
     "single-page-app"
@@ -28,15 +28,19 @@
   },
   "homepage": "https://github.com/Lucifier129/create-app#readme",
   "devDependencies": {
-    "babel": "^5.8.35",
-    "babel-core": "^5.8.25",
-    "babel-jest": "^5.3.0",
-    "babel-loader": "^5.3.2",
-    "babel-runtime": "^5.8.25",
+    "babel-cli": "^6.14.0",
+    "babel-loader": "^6.2.5",
+    "babel-polyfill": "^6.13.0",
+    "babel-preset-latest": "^6.14.0",
+    "babel-preset-react": "^6.11.1",
+    "babel-preset-stage-0": "^6.5.0",
+    "babel-register": "^6.14.0",
     "bundle-loader": "^0.5.4",
     "expect": "^1.20.2",
+    "fastclick": "^1.0.6",
     "gulp": "^3.9.1",
     "gulp-gh-pages": "^0.5.4",
+    "isomorphic-fetch": "^2.2.1",
     "karma": "^1.1.2",
     "karma-chrome-launcher": "^1.0.1",
     "karma-mocha": "^1.1.1",
@@ -44,13 +48,10 @@
     "karma-sourcemap-loader": "^0.3.7",
     "karma-webpack": "^1.7.0",
     "mocha": "^2.5.3",
-    "react": "^15.2.1",
-    "react-dom": "^15.2.1",
-    "react-lite": "^0.15.14",
-    "rollup": "^0.21.0",
-    "rollup-plugin-babel": "^1.0.0",
-    "rollup-plugin-replace": "^1.1.0",
-    "uglify-js": "^2.6.1",
+    "react": "^15.3.1",
+    "react-dom": "^15.3.1",
+    "react-lite": "^0.15.17",
+    "relite": "^1.0.0",
     "webpack": "^1.12.2"
   },
   "dependencies": {
