@@ -2,16 +2,11 @@ import createApp from 'create-app/client'
 import ReactDOM from 'react-dom'
 import routes from './routes'
 import config from './config'
-import 'isomorphic-fetch'
-import * as History from 'history'
 
-const webpackLoader = (url, initController) => {
-    var load = require(url)
-    load(module => {
-        let Controller = module.default || module
-        return initController(Controller)
-    })
-}
+const webpackLoader = (url) => (
+    new Promise(require(url))
+    .then(module => module.default || module)
+)
 
 const viewEngine = {
     render: (component, container) => {
