@@ -1,15 +1,13 @@
-// import server from './server/index'
 'use strict';
 
-Object.defineProperty(exports, '__esModule', {
+Object.defineProperty(exports, "__esModule", {
   value: true
 });
 
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
+var _index = require('./client/index');
 
-var _clientIndex = require('./client/index');
+var _index2 = _interopRequireDefault(_index);
 
-var _clientIndex2 = _interopRequireDefault(_clientIndex);
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-exports['default'] = _clientIndex2['default'];
-module.exports = exports['default'];
+exports.default = _index2.default; // import server from './server/index'
