@@ -1,13 +1,12 @@
-/**
- * default view engine for server
- */
-
 "use strict";
 
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
-var render = function render(html) {
+/**
+ * default view engine for server
+ */
+
+var render = exports.render = function render(html) {
   return html.toString();
 };
-exports.render = render;
