@@ -1,19 +1,16 @@
+'use strict';
+
+Object.defineProperty(exports, "__esModule", {
+	value: true
+});
 /*
   key/value configs
 */
-'use strict';
-
-Object.defineProperty(exports, '__esModule', {
-	value: true
-});
-var isClient = typeof window !== 'undefined';
-exports.isClient = isClient;
-var isServer = !isClient;
-exports.isServer = isServer;
-var defaultAppSettings = {
+var isClient = exports.isClient = typeof window !== 'undefined';
+var isServer = exports.isServer = !isClient;
+var defaultAppSettings = exports.defaultAppSettings = {
 	container: '#container',
 	basename: '',
 	context: {},
 	type: 'createHistory'
 };
-exports.defaultAppSettings = defaultAppSettings;
