@@ -1,13 +1,9 @@
 'use strict';
 
-Object.defineProperty(exports, '__esModule', {
+Object.defineProperty(exports, "__esModule", {
     value: true
 });
-exports['default'] = createMatcher;
-
-function _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj; } else { var newObj = {}; if (obj != null) { for (var key in obj) { if (Object.prototype.hasOwnProperty.call(obj, key)) newObj[key] = obj[key]; } } newObj['default'] = obj; return newObj; } }
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
+exports.default = createMatcher;
 
 var _pathToRegexp = require('path-to-regexp');
 
@@ -16,6 +12,10 @@ var _pathToRegexp2 = _interopRequireDefault(_pathToRegexp);
 var _util = require('./util');
 
 var _ = _interopRequireWildcard(_util);
+
+function _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj; } else { var newObj = {}; if (obj != null) { for (var key in obj) { if (Object.prototype.hasOwnProperty.call(obj, key)) newObj[key] = obj[key]; } } newObj.default = obj; return newObj; } }
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 function createMatcher(routes) {
     var finalRoutes = routes.map(createRoute);
@@ -43,7 +43,7 @@ function createMatcher(routes) {
 function createRoute(route) {
     var finalRoute = _.extend({}, route);
     var keys = finalRoute.keys = [];
-    finalRoute.regexp = (0, _pathToRegexp2['default'])(finalRoute.path, keys);
+    finalRoute.regexp = (0, _pathToRegexp2.default)(finalRoute.path, keys);
     return finalRoute;
 }
 
@@ -65,4 +65,3 @@ function getParams(matches, keys) {
 function cleanPath(path) {
     return path.replace(/\/\//g, '/');
 }
-module.exports = exports['default'];
