@@ -1,49 +1,49 @@
+'use strict';
+
+Object.defineProperty(exports, "__esModule", {
+	value: true
+});
+
+var _createBrowserHistory = require('create-history/lib/createBrowserHistory');
+
+var _createBrowserHistory2 = _interopRequireDefault(_createBrowserHistory);
+
+var _createHashHistory = require('create-history/lib/createHashHistory');
+
+var _createHashHistory2 = _interopRequireDefault(_createHashHistory);
+
+var _useQueries = require('create-history/lib/useQueries');
+
+var _useQueries2 = _interopRequireDefault(_useQueries);
+
+var _useBeforeUnload = require('create-history/lib/useBeforeUnload');
+
+var _useBeforeUnload2 = _interopRequireDefault(_useBeforeUnload);
+
+var _useBasename = require('create-history/lib/useBasename');
+
+var _useBasename2 = _interopRequireDefault(_useBasename);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+// beforeunload support
+
+
+// Hash history
+exports.default = {
+	createHistory: _createBrowserHistory2.default,
+	createHashHistory: _createHashHistory2.default,
+	useQueries: _useQueries2.default,
+	useBeforeUnload: _useBeforeUnload2.default,
+	useBasename: _useBasename2.default
+};
+
+// basename support
+
+
+// query support
 /**
  * custom history import
  */
 
 // HTML5 history
-'use strict';
-
-Object.defineProperty(exports, '__esModule', {
-	value: true
-});
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
-
-var _createHistoryLibCreateBrowserHistory = require('create-history/lib/createBrowserHistory');
-
-var _createHistoryLibCreateBrowserHistory2 = _interopRequireDefault(_createHistoryLibCreateBrowserHistory);
-
-// Hash history
-
-var _createHistoryLibCreateHashHistory = require('create-history/lib/createHashHistory');
-
-var _createHistoryLibCreateHashHistory2 = _interopRequireDefault(_createHistoryLibCreateHashHistory);
-
-// query support
-
-var _createHistoryLibUseQueries = require('create-history/lib/useQueries');
-
-var _createHistoryLibUseQueries2 = _interopRequireDefault(_createHistoryLibUseQueries);
-
-// beforeunload support
-
-var _createHistoryLibUseBeforeUnload = require('create-history/lib/useBeforeUnload');
-
-var _createHistoryLibUseBeforeUnload2 = _interopRequireDefault(_createHistoryLibUseBeforeUnload);
-
-// basename support
-
-var _createHistoryLibUseBasename = require('create-history/lib/useBasename');
-
-var _createHistoryLibUseBasename2 = _interopRequireDefault(_createHistoryLibUseBasename);
-
-exports['default'] = {
-	createHistory: _createHistoryLibCreateBrowserHistory2['default'],
-	createHashHistory: _createHistoryLibCreateHashHistory2['default'],
-	useQueries: _createHistoryLibUseQueries2['default'],
-	useBeforeUnload: _createHistoryLibUseBeforeUnload2['default'],
-	useBasename: _createHistoryLibUseBasename2['default']
-};
-module.exports = exports['default'];
