@@ -1,13 +1,12 @@
-// util
 'use strict';
 
-Object.defineProperty(exports, '__esModule', {
+Object.defineProperty(exports, "__esModule", {
     value: true
 });
 exports.isThenable = isThenable;
 exports.identity = identity;
 exports.extend = extend;
-
+// util
 function isThenable(obj) {
     return obj != null && typeof obj.then === 'function';
 }
